backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.4.2
alembic==1.12.1
psycopg2-binary==2.9.9